Encrypt Decrypt any of the file stored in your computer with user defined key
Python based project that takes the file path as input and a key, then
it encrypts it with SHA256 algo. The original file is replaced with the
encrypted file.
If you wish to decrypt it, take the file input again and input the key
you entered earlier.
The file will be decrypted back.